delete zebra/slankdev_netlink.h
